Update sample apps for encoding OC routing policy
Includes one boilerplate app and four custom apps to encode OC routing
policy configuration in XML:
cd-encode-oc-routing-policy-10-ydk.py - boilerplate
cd-encode-oc-routing-policy-22-ydk.py - accept all prefixes
cd-encode-oc-routing-policy-24-ydk.py - match path/comm, set comm/lpref
cd-encode-oc-routing-policy-26-ydk.py - match prefix, set comm/lpref
cd-encode-oc-routing-policy-28-ydk.py - set next hop self
Refactors previous sample apps to make them compatible with model
bundles introduced in YDK 0.5.0.  App names now use a notation that is
more service and model agnostic.  Previous apps were committed in
https://github.com/CiscoDevNet/ydk-py-samples/commit/b46e37a87d73ded101a6c27988c99ef58dd3d344
