+++ samples/basic/codec/models/openconfig/openconfig-routing-policy/cd-encode-oc-routing-policy-22-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model openconfig-routing-policy.
+Encode configuration for model openconfig-routing-policy.
 
-usage: cd-encode-config-routing-policy-28-ydk.py [-h] [-v]
+usage: cd-encode-oc-routing-policy-22-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,21 +30,19 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.routing import routing_policy as oc_routing_policy
-from ydk.models.bgp import bgp_policy as oc_bgp_policy
+from ydk.models.openconfig import openconfig_routing_policy \
+    as oc_routing_policy
 from ydk.types import Empty
 import logging
 
 
 def config_routing_policy(routing_policy):
     """Add config data to routing_policy object."""
-    # configure policy definition
+    # configure routing policy
     policy_definition = routing_policy.policy_definitions.PolicyDefinition()
-    policy_definition.name = "POLICY4"
+    policy_definition.name = "POLICY1"
     statement = policy_definition.statements.Statement()
-    statement.name = "next-hop-self"
-    set_next_hop = oc_bgp_policy.BgpNextHopTypeEnum.SELF
-    statement.actions.bgp_actions.set_next_hop = set_next_hop
+    statement.name = "accept route"
     statement.actions.accept_route = Empty()
     policy_definition.statements.statement.append(statement)
     routing_policy.policy_definitions.policy_definition.append(policy_definition)
@@ -73,10 +71,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    routing_policy = oc_routing_policy.RoutingPolicy()  # create config object
+    routing_policy = oc_routing_policy.RoutingPolicy()  # create object
     config_routing_policy(routing_policy)  # add object configuration
 
-    print(codec.encode(provider, routing_policy))  # encode and print object
+    # encode and print object
+    print(codec.encode(provider, routing_policy))
+
     provider.close()
     exit()
 # End of script
