+++ samples/basic/codec/models/openconfig/openconfig-routing-policy/cd-encode-oc-routing-policy-22-ydk.xml
