+++ samples/basic/codec/models/openconfig/openconfig-routing-policy/cd-encode-oc-routing-policy-24-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model openconfig-routing-policy.
+Encode configuration for model openconfig-routing-policy.
 
-usage: cd-encode-config-routing-policy-24-ydk.py [-h] [-v]
+usage: cd-encode-oc-routing-policy-24-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,7 +30,8 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.routing import routing_policy as oc_routing_policy
+from ydk.models.openconfig import openconfig_routing_policy \
+    as oc_routing_policy
 from ydk.types import Empty
 import logging
 
@@ -110,10 +111,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    routing_policy = oc_routing_policy.RoutingPolicy()  # create config object
+    routing_policy = oc_routing_policy.RoutingPolicy()  # create object
     config_routing_policy(routing_policy)  # add object configuration
 
-    print(codec.encode(provider, routing_policy))  # encode and print object
+    # encode and print object
+    print(codec.encode(provider, routing_policy))
+
     provider.close()
     exit()
 # End of script
