+++ samples/basic/codec/models/openconfig/openconfig-routing-policy/cd-encode-oc-routing-policy-24-ydk.xml
