+++ samples/basic/codec/models/openconfig/openconfig-routing-policy/cd-encode-oc-routing-policy-26-ydk.py
@@ -16,9 +16,9 @@
 #
 
 """
-Encode config for model openconfig-routing-policy.
+Encode configuration for model openconfig-routing-policy.
 
-usage: cd-encode-config-routing-policy-22-ydk.py [-h] [-v]
+usage: cd-encode-oc-routing-policy-26-ydk.py [-h] [-v]
 
 optional arguments:
   -h, --help     show this help message and exit
@@ -30,20 +30,57 @@ from urlparse import urlparse
 
 from ydk.services import CodecService
 from ydk.providers import CodecServiceProvider
-from ydk.models.routing import routing_policy as oc_routing_policy
+from ydk.models.openconfig import openconfig_routing_policy \
+    as oc_routing_policy
 from ydk.types import Empty
 import logging
 
 
 def config_routing_policy(routing_policy):
     """Add config data to routing_policy object."""
-    # configure routing policy
+    # configure prefix set
+    prefix_set = routing_policy.defined_sets.prefix_sets.PrefixSet()
+    prefix_set.prefix_set_name = "PREFIX-SET1"
+    prefix = prefix_set.Prefix()
+    prefix.ip_prefix = "10.0.0.0/16"
+    prefix.masklength_range = "24..32"
+    prefix_set.prefix.append(prefix)
+    prefix = prefix_set.Prefix()
+    prefix.ip_prefix = "172.0.0.0/8"
+    prefix.masklength_range = "16..32"
+    prefix_set.prefix.append(prefix)
+    routing_policy.defined_sets.prefix_sets.prefix_set.append(prefix_set)
+
+    # configure community set
+    bgp_defined_sets = routing_policy.defined_sets.bgp_defined_sets
+    community_set = bgp_defined_sets.community_sets.CommunitySet()
+    community_set.community_set_name = "COMMUNITY-SET2"
+    community_set.community_member.append("65172:17001")
+    bgp_defined_sets.community_sets.community_set.append(community_set)
+
+    # configure policy definition
     policy_definition = routing_policy.policy_definitions.PolicyDefinition()
-    policy_definition.name = "POLICY1"
+    policy_definition.name = "POLICY3"
+    # prefix-set statement
     statement = policy_definition.statements.Statement()
-    statement.name = "accept route"
+    statement.name = "prefix-set1"
+    match_prefix_set = statement.conditions.MatchPrefixSet()
+    match_prefix_set.prefix_set = "PREFIX-SET1"
+    match_set_options = oc_routing_policy.MatchSetOptionsRestrictedTypeEnum.ANY
+    match_prefix_set.match_set_options = match_set_options
+    statement.conditions.match_prefix_set = match_prefix_set
+    statement.actions.bgp_actions.set_local_pref = 1000
+    set_community = statement.actions.bgp_actions.SetCommunity()
+    set_community.community_set_ref = "COMMUNITY-SET2"
+    statement.actions.bgp_actions.set_community = set_community
     statement.actions.accept_route = Empty()
     policy_definition.statements.statement.append(statement)
+    # reject statement
+    statement = policy_definition.statements.Statement()
+    statement.name = "reject route"
+    statement.actions.reject_route = Empty()
+    policy_definition.statements.statement.append(statement)
+
     routing_policy.policy_definitions.policy_definition.append(policy_definition)
 
 
@@ -70,10 +107,12 @@ if __name__ == "__main__":
     # create codec service
     codec = CodecService()
 
-    routing_policy = oc_routing_policy.RoutingPolicy()  # create config object
+    routing_policy = oc_routing_policy.RoutingPolicy()  # create object
     config_routing_policy(routing_policy)  # add object configuration
 
-    print(codec.encode(provider, routing_policy))  # encode and print object
+    # encode and print object
+    print(codec.encode(provider, routing_policy))
+
     provider.close()
     exit()
 # End of script
