--- samples/basic/codec/ydk/models/openconfig/cd-encode-config-routing-policy-22-ydk.txt
@@ -1,6 +1,0 @@
-route-policy POLICY1
-  #statement-name accept route
-  done
-end-policy
-!
-end
