--- samples/basic/codec/ydk/models/openconfig/cd-encode-config-routing-policy-26-ydk.txt
@@ -1,21 +1,0 @@
-prefix-set PREFIX-SET1
-  10.0.0.0/16 ge 24 le 32,
-  172.0.0.0/8 ge 16 le 32
-end-set
-!
-community-set COMMUNITY-SET2
-  65172:17001
-end-set
-!
-route-policy POLICY3
-  #statement-name prefix-set1
-  if destination in PREFIX-SET1 then
-    set local-preference 1000
-    set community COMMUNITY-SET2
-    done
-  endif
-  #statement-name reject
-  drop
-end-policy
-!
-end
