--- samples/basic/codec/ydk/models/openconfig/cd-encode-config-routing-policy-28-ydk.txt
@@ -1,7 +1,0 @@
-route-policy POLICY4
-  #statement-name next-hop-self
-  set next-hop self
-  done
-end-policy
-!
-end
